refactor(shared): split schema.ts under 800-line cap (#460)
Trunk left schema.ts at 798 (2 under cap); the add-on table + addOnSnapshot
column/payload field push it over. Extract the add_on_options table+enum to
db/add-on.ts and the booking snapshot/event payload types to db/booking-types.ts,
both re-exported from schema.ts so drizzle-kit (loads only this module) and
existing importers are unaffected. No SQL change (db:verify: no drift). schema.ts
now 772. Also fill required addOnSnapshot:[] in trunk's seed + payment fixtures
(merge made the field required).
