+++ packages/api/src/services/payment/payment.test.ts
@@ -35,6 +35,7 @@ function makeBooking(overrides: Partial<Booking> = {}): Booking {
     insuranceOptionId: null,
     insuranceSnapshot: null,
     feeSnapshot: [],
+    addOnSnapshot: [],
     externalId: null,
     notes: null,
     totalPrice: 100_000,
